added normalize utility for the call function
